refactor(732): extract codegen-free shared monomorphizer
Move generic-instantiation discovery + AST substitution out of
vera/codegen/monomorphize.py into a new codegen-free module
vera/monomorphize.py (MonoContext + Monomorphizer) so the verifier
(#732) can reuse the EXACT same discovery codegen uses.  This is a
soundness requirement: per-monomorphization static verification is only
sound if the verifier checks precisely the instantiation set codegen
emits — a divergent second inference path could miss an instantiation
and produce a false Tier-1.

- vera/monomorphize.py (new): leaf inference + substitution
  (_collect_calls_in_expr, _infer_type_args_from_call, _monomorphize_fn,
  _substitute_in_ast, De Bruijn reindex, ...) over a MonoContext of
  registration metadata.  Imports only vera.ast.  substitute_type_vars
  relocated here (was vera/wasm/inference.py) so importing the
  monomorphizer doesn't drag in the WASM backend; re-exported from
  vera/wasm/inference.py for its four other importers.
- vera/codegen/monomorphize.py: MonomorphizationMixin builds a
  MonoContext and delegates leaf work to Monomorphizer, keeping its
  orchestration (seed walk + transitive worklist with the
  constraint-failing filter) and the layout-specific _check_constraints
  / _adt_satisfies_eq (E613) exactly as before.  fn_ret_types
  reproduces the old _fn_sigs i64->Int/i32->Bool/f64->Float64 mapping
  so _infer_fncall_vera_type_simple stays byte-identical.

Pure refactor, no behavior change.  Gate: codegen + wasm + monomorphize
suites pass (byte-identical WAT), full suite 4519 passed / 15 skipped,
mypy + ruff + ruff --select S clean.
